requirements.txt
asgiref==3.11.0
attrs==25.4.0
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
dj-database-url==2.1.0
Django==5.2.10
django-admin-interface==0.32.0
django-colorfield==0.14.0
django-cors-headers==4.6.0
django-extensions==4.1
django-filter==24.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
djlint==1.36.4
drf-spectacular==0.27.2
EditorConfig==0.17.1
gunicorn==25.0.1
inflection==0.5.1
jsbeautifier==1.15.4
json5==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pathspec==1.0.3
pillow==12.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.11.0